Major clean-up of old leaf-splitting code.
Summary: the functions

   leaftools.split( )
   leaftools.split_binary( )

have been condensed into a single functinon:

   leaftools.split_general( )

In addition, the functions

   leaftools.scale( )
   leaftools.scale_binary( )

have been completely removed in favor of the function

   leaftools.duration_change( )

which has been sitting around in the codebase as a competing pattern that generalizes both leaftools.scale( ) and leaftools.scale_binary( ).

The old leaftools.split( ) and leaftools.split_binary( ) have been in the system forever, each with an annoying set of idiosyncracies that made it difficult to generalize the functions. Among the annoyances were the following:

1. Both functions spliced newly created leaves TO THE LEFT of the input leaf. This is odd. In general, the pattern should be that newly created material splices TO THE RIGHT of input material. (The only exceptions to the this should be clearly marked in the name of the function with something like 'backwards' or 'rightwards' or something similar.)

2. The two functions returned different types of structure. Specifically, leaftools.split_binary( ) returned a left list and a right list. The left list comprised either a single untied note or else two or more tied notes. The right list comprised the same. Compare this with leaftools.split( ) which returned simply two Abjad components, usually FixedDurationTuplets. This situation made it difficult to type- and structure-check the output of the two functions in a general way.

3. The functions duplicated quite a bit of code. In general, leaftools.split( ) was a wrapper around leaftools.scale( ) and leaftools.split_binary( ) was a wrapper around leaftools.scale_binary( ). But both functions included a dozen or so lines that were absolutely identical. These lines did things like handled after graces and before graces and articulations during split. This code has survived and migrated into a single place, that is, in leaftools.split_general( ).

The new leaftools.split_general( ) answers these concerns as follows:

1. The new function splices only the right.

2. The new function returns a list with left and right parts only.

3. The new function duplicates no code.

Examples:

In general there are three different rhythmic breakpoints at which you might want to split a leaf. It was our failure to perceive this three-part pattern that lead to the inelegancies in the old functions. The three different types of rhythmic breakpoints for leaf splitting are these:

A. At a notehead-assignable point, such as 1/16th into quarter note. This type of split requires neither ties nor tuplets.

B. At a point that is not notehead-assignable but does still have a binary denominator, such as 5/64ths into a quarter note. This type of split requires a tie but no tuplets.

C. At a point that is not only not notehead-assignable but also carries a nonbinary denominator, such as 1/12th into a quarter note. This type of split requires a tuplet to be created from thin air.

If we want to go further we can add a fourth type of rhythmic breakpoint:

D. At a point that requires both ties and a de novo tuplet, such as 5/48 into a quarter note.


Here are examples of each of the four types of rhythmic breakpoint listed above, all within a container and spanned in a realworld type of way:

Starting structure:

   t = Staff(construct.scale(3))
   Beam(t[:])

   \new Staff {
           c'8 [
           d'8
           e'8 ]
   }


Example A, with neither ties nor tuplets:

   leaftools.split_general(t[1], Rational(1, 32))

   \new Staff {
           c'8 [
           d'32
           d'16.
           e'8 ]
   }


Example B, with ties but without tuplets:

   \new Staff {
           c'8 [
           d'16 ~
           d'64
           d'32.
           e'8 ]
   }


Example C, without ties but with a tuplet:

   \new Staff {
           c'8 [
           \times 2/3 {
                   d'8
           }
           \times 2/3 {
                   d'16
           }
           e'8 ]
   }


Example D, with ties and with a tuplet:

   \new Staff {
           c'8 [
           \times 2/3 {
                   d'8 ~
                   d'32
           }
           \times 2/3 {
                   d'32
           }
           e'8 ]
   }


This is the new test file:

   tools/leaftools/test/test_leaftools_split_general.py

The test file contains 16 total examples.

Note too that this commit preserves a great deal of the previous code in comments and in skipped test files. The next commit will remove most of this.