External measuretools.trim_by_count( ) moved to RigidMeasure.__delitem__( ).
The external function was, contrary to its name, acting on a Python list of component references, rather than on a number of components. This suggested moving the functionality to RigidMeasure.__delitem__( ) which must operate on a Python list of component references, all of which are measure contents.

This means that Python del now works as expected on rigid measures.

Example:

   abjad> t = RigidMeasure((4, 8), construct.scale(4))
   abjad> Beam(t[:])
   Beam(c'8, d'8, e'8, f'8)

   abjad> del(t[:2])

   abjad> f(t)
           \time 2/8
           e'8 [
           f'8 ]