Pair of new public functions added to tools/partition package:
   partition.fractured_by_durations(components, durations)
   partition.unfractured_by_durations(components, durations)


Current package contents:

   partition.container_cyclic_fractured_by_counts( )
   partition.container_cyclic_unfractured_by_counts( )
   partition.container_fractured_by_counts( )
   partition.container_unfractured_by_counts( )
   partition.fractured_by_durations( )
   partition.unfractured_by_durations( )
   partition.leaves_cyclic_unfractured_by_durations( )


The new functions take a list of zero or more Abjad components and a list of zero or more rational durations. Components in list are then split at exactly the points equal to the cumulative sum of the durations in list. Spanners are either fractured or not fractured, according to the name of the new function used.

Example:

Consider the following score fragment:

   t = Staff(Container(construct.run(2)) * 2)
   pitchtools.diatonicize(t)
   Beam(t[0])
   Beam(t[1])
   Slur(t.leaves)

   r'''\new Staff {
      {
         c'8 [ (
         d'8 ]
      }
      {
         e'8 [
         f'8 ] )
      }
   }'''

Now partition both containers in score and fracture spanners attached directly to the two containers:

   durations = [Rational(1, 32), Rational(3, 32), Rational(5, 32)]
   parts = partition.fractured_by_durations(t[:], durations)

   r'''\new Staff {
      {
         c'32 [ ] (
      }
      {
         c'16. [ ]
      }
      {
         d'8 [ ]
      }
      {
         e'32 [ ]
      }
      {
         e'16. [
         f'8 ] )
      }
   }'''


Example:

Consider the same score fragment again. Then partition both containers in score but leave spanners untouched.

   durations = [Rational(1, 32), Rational(3, 32), Rational(5, 32)]
   parts = partition.unfractured_by_durations(t[:], durations)

   r'''\new Staff {
      {
         c'32 [ (
      }
      {
         c'16.
      }
      {
         d'8 ]
      }
      {
         e'32 [
      }
      {
         e'16.
         f'8 ] )
      }
   }'''


The functions are powerfully generalized to work on all Abjad components, including leaves, tuplets, measures, contexts and unqualified containers. Further, this new pair of functions is generalized to work on component both outside of score and inside score. That is, you can freely partition arbitrary Python lists of Abjad components not yet incorporated in score.

The test files contain additional examples.