Extended leaf split logic to optionally fracture spanners.
Previously there had been a gap in the system such that container split logic optionally fractured spanners but that leaf split logic never fractured spanners. The reason for this was simply that the leaf split logic was implemented months before most of the container split logic.

This revision extends the leaf split logic and brings it in line with the container split logic.

Example 1. Duration split leaf in score and fracture spanners:

   t = Staff(construct.scale(4))
   Beam(t[:])

   \new Staff {
           c'8 [
           d'8
           e'8
           f'8 ]
   }

   >>> split.fractured_at_duration(t.leaves[2], Rational(1, 32))

   \new Staff {
           c'8 [
           d'8
           e'32 ]
           e'16. [
           f'8 ]
   }


Example 2. Duration partition leaf in score and fracture spanners:

   t = Staff(construct.scale(4))
   Beam(t[:])

   \new Staff {
           c'8 [
           d'8
           e'8
           f'8 ]
   }

   >>> durations = [Rational(1, 64), Rational(5, 64)]
   >>> partition.fractured_by_durations(t[2:3], durations)

   \new Staff {
           c'8 [
           d'8
           e'64 ]
           e'16 [ ~
           e'64 ]
           e'32 [
           f'8 ]
   }

The test files contain many examples.