Important modification to both the *behavior* and *syntax* of `abjad-book`.
** Syntax:
The opening and closing <abjad> </abjad> tags are kept.
The internal `abjad>`, `>>>` and `hide>` are now deprecated. It is no longer necessary to insert any of these tags at the beginning of each line in an <abjad></abjad> block. Simply write each line as you would in Python. 
To hide a line there is now the `<hide` internal tag. Notice that this is just like the old `hide>`, but with the arrow pointing back. This is because the tag is intended to be a suffix tag. i.e. go at the end, rather than at the beginning of the line. This allows one to keep the indentation (fundamental in Python) of the code clear.

Currently, all the code inside the <abjad> </abjad> block *must* be fully ragged to the left. i.e. absolutely no indentation must exist in the first line of the code block. This may be changed in the future to make the formatting of the raw document more flexible. 

** Behavior:
`abjad-book` now captures *all* Python output generated during the execution of code blocks. This means that it is no longer necessary to explicitly write formatting output by hand in, for example, `f(score)`.
Example:

<abjad>
v = Voice(construct.run(8))
print v
f(v)
</abjad>


results in the following output:


v = Voice(construct.run(8))
print v
Voice{8}
f(v)
\new Voice {
   c'8
   c'8
   c'8
   c'8
   c'8
   c'8
   c'8
   c'8
}