Tempo interface now respects spanners at tempo-determination time.
To make this work it was necessary to come up with at least a provisional solution for what happens when a single leaf is both spanned by a tempo spanner and is also the carrier of a forced tempo indication. 

The solution implemented here is easiest to see in the new definition of ``_TempoInterface.effective``::

   @property
   def effective(self):
      '''Effective tempo governing client.
         Decisions here arbitrate between spanner and forced attribute.'''
      if self.forced:
         return self.forced
      if self.spanned:
         return self.spanner.indication
      return _BacktrackingInterface.effective.fget(self)

That is, if the leaf carries a forced tempo indication, return the forced tempo indication as the effective tempo governing the leaf in question. Otherwise, if the leaf is spanned with a tempo spanner, return the tempo indication of the spanner that governs the leaf. Otherwise, backtrack and find the last best forced tempo indication and return that as the effective tempo of the leaf in question.

This situation isn't stable and it probably makes more sense to insist on setting tempo with spanners only, just as we've done with ties and glissandi previously.

Tests in ``tempo/test/test_tempo_spanner.py`` now pass.