New ``spanner_in_parentage`` attribute implemented on ``_SpannerReceptor`` class. The attribute is defined to return the first spanner of appropriate class in the parentage of the client of the spanner receptor in question. Compare with the well established ``spanner`` attribute already implemented on ``_SpannerReceptor``::
   t = Voice(RigidMeasure((2, 8), construct.run(2)) * 2)
   p = Tempo(t[:], TempoIndication(Rational(1, 8), 38))

   \new Voice {
         \time 2/8
         \tempo 8=38
         c'8
         c'8
         \time 2/8
         c'8
         c'8
         %% tempo 8=38 ends here
   }

   >>> t.leaves[0].tempo.spanner
   Traceback (most recent call last):
     File "<stdin>", line 1, in <module>
     File "/Users/trevorbaca/Documents/abjad/trunk/abjad/spanner/receptor.py", line 82, in spanner
       raise MissingSpannerError
   abjad.exceptions.exceptions.MissingSpannerError

   >>> t.leaves[0].tempo.spanner_in_parentage
   Tempo(|2/8, c'8, c'8|, |2/8, c'8, c'8|)

This new implementation of ``spanner_in_parentage`` returns to the still-open question of the best way to examine spanners attaching to the parentage (or possibly children) of a given component. The current implementation of ``_SpannerReceptor`` was originally designed to serve information about only those spanners attaching **directly** to the client of the spanner receptor only. However, both the existing ``parented`` and new ``spanner_in_parentage`` attributes implemented on ``_SpannerReceptor`` serve information about spanners attaching to components in the parentage of client, rather than directly to client itself.