New ``_SpacingInterface`` and ``SpacingSpanner``.
``_SpacingInterface`` and ``SpacingSpanner`` both handle the *LilyPond* ``SpacingSpanner`` grob. In addition, the *Abjad* ``SpacingSpanner`` implements a read / write ``new_section`` attribute as an interface to the *LilyPond* ``newSpacingSection`` command.

Remember that the *LilyPond* ``SpacingSpanner`` grob usually lives at the level of the *LilyPond* ``Score`` context. So ``SpacingSpanner`` grob overrides in *Abjad* make the most sense at the level of the *Abjad* ``Score``.

Example. Overriding the *LilyPond* ``SpacingSpanner`` grob at the *Abjad* ``Score``::

   t = Score([ ])
   t.spacing.strict_grace_spacing = True
   t.spacing.strict_note_spacing = True
   t.spacing.uniform_stretching = True

   \new Score \with {
           \override SpacingSpanner #'strict-note-spacing = ##t
           \override SpacingSpanner #'uniform-stretching = ##t
           \override SpacingSpanner #'strict-grace-spacing = ##t
   } <<
   >>

Example. Applying *LilyPond* spacing section to *Abjad* leaves::

   t = Staff(construct.scale(4))
   p = SpacingSpanner(t[:])
   p.new_section = True

   \new Staff {
           \newSpacingSection
           c'8
           d'8
           e'8
           f'8
           %%% spacing section ends here %%%
   }

These examples are included in the test files attaching to this commit.