New read-only ``explicit`` attribute added to ``_StaffInterface``.
You can now say ``t.staff.explicit`` for any *Abjad* component ``t`` that has a ``_StaffInterface``. The attribute returns a reference to the first explicit *Abjad* ``Staff`` in the parentage of ``t``.

Note that only leaves and measures currently contain a ``_StaffInterface``. Voice, staves, scores and other types of container do not currently contain a ``_StaffInterface``.