New ``TempoProportional`` spanner.
The new ``TempoProportional`` spanner is a smart spanner in something of the same style as the ``BeamComplex`` spanner.

The ``TempoProportional`` spanner extends the naive ``Tempo`` spanner with logic to apply new spacing sections to the spanner contents and to space spanner accordings according to score-global spacing (if score-global spacing is defined).

Example::

   t = Score([Staff(construct.scale(4))])
   tempo_indication = TempoIndication(Rational(1, 4), 60)
   spacing_indication = SpacingIndication(tempo_indication, Rational(1, 34))
   t.global_spacing = spacing_indication

   p = TempoProportional(t[0][:2])
   p.indication = TempoIndication(Rational(1, 4), 60)
   p = TempoProportional(t[0][2:])
   p.indication = TempoIndication(Rational(1, 4), 120)

   \new Score <<
           \new Staff {
                   \tempo 4=60
                   \newSpacingSection
                   \set Score.proportionalNotationDuration = #(ly:make-moment 1 . 34)
                   c'8
                   d'8
                   %% tempo 4=60 ends here
                   \tempo 4=120
                   \newSpacingSection
                   \set Score.proportionalNotationDuration = #(ly:make-moment 1 . 17)
                   \tempo 4=120
                   e'8
                   f'8
                   %% tempo 4=120 ends here
           }
   >>

What's going on here is that score-global spacing is defined for the score given above. That is, score-global spacing is set to space according to 1/34 whenever local tempo is 4=60. Following this definition of score-global spacing, tempo different proportional tempo spanners apply to the four leaves in the score. The first proportional tempo spanner defines tempo of 4=60, which just happens to equal the score-global tempo. The second proportional tempo spanner defines a tempo of 4=120, which is exactly twice as fast as the score-global tempo of 4=60. At format-time, both proportional tempo spanners define new spacing sections and set the *LilyPond* ``proportionalNotationDuration`` setting in the score context to values that calculated against score-global spacing.