Elaborated docstrings in ``voice/interface/interface.py`` module.
Docstrings for this module are now complete. Each attribute in the ``_VoiceInterface`` public interface:

   * appears in the Sphinx API
   * carries a 1-sentence description
   * gives sample usage
   * gives read / write status
   * lists important additional pieces of information

Also, it turns out that Sphinx respects backslash as the Python line-continuation character within docstrings. So it is still possible to respect the 80-character max width part of the Abjad coding standards while elaborating docstrings for Sphinx.