New ``listtools.partition_by_counts( )`` function.
Signature::

   partition_by_counts(l, s, cyclic = False, overhang = False)

Example::

   abjad> l = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
   abjad> listtools.partition_by_counts(l, [3])
   [[0, 1, 2]]

When ``cyclic = True`` repeat the elements in ``s``.

::

   abjad> l = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
   abjad> listtools.partition_by_counts(l, [3], cyclic = True) 
   [[0, 1, 2], [3, 4, 5], [6, 7, 8]]

When ``overhang = True`` return any remaining unicorporated \
elements of ``l`` as a final part in ``result``.

      abjad> l = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
      abjad> listtools.partition_by_counts(l, [3], overhang = True)
      [[0, 1, 2], [3, 4, 5, 6, 7, 8, 9]]

When both ``cyclic = True`` and ``overhang = True`` repeat the \
elements in ``s`` and return any remaining unincorporated \
elements of ``l`` as a final part in ``result``.

::

   abjad> l = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
   abjad> listtools.partition_by_counts(
      l, [3], cyclic = True, overhang = True)
   [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9]]

See the API doc chapter for even more examples.