Changed ``listtools.truncate_to_sum(l, total)`` at boundary case when ``total == 0``.
Previously ``listtools.truncate_to_sum(l, 0)`` returned a new one-element list ``[0]`` containing the number ``0``, regardless of the contents of *l*.

Now ``listtools.truncate_to_sum(l, 0)`` returns a new empty list, again regardless of the contents of *l*.

The previous behavior can not have been correct. For surely it can not be right to truncate [1, 2, 3, 4, 5] to [0] given that ``0`` exists nowhere in the input list.

Changed the test case that asserted the previous behavior, too.