``io.write_pdf( )``  enhanced:
The function now uses the cross-platform `shutil` Python package to move the rendered PDF file, rather than the POSIX specific ``os.system('mv')``. 
The function now also allows writing the PDF file anywhere in the system. The given `file_name` is expanded (in case ``~`` is present in the file name) and ``.pdf`` is appended to `file_name` if not already present.