Added active class package initializers.
This commit starts with the abjad/skip package.

The core typesetting classes -- Note, Rest, Chord, Skip, Staff, Voice, and so on -- implemented in previous versions of the system have what we might call 'null' or 'inactive' or 'passive' initializers. That is, note/__init__.py, rest/__init__.py and so on are all, in previous versions of the system, empty files. These 'null' initializes serve the useful purpose of marking the note/, rest/, chord/ directories as packages, which makes it possible to structure class imports in, for example, the following way::

   chord/initializer.py:23:            from abjad.skip.skip import Skip
   metricgrid/format.py:3:             from abjad.skip.skip import Skip
   note/initializer.py:12:             from abjad.skip.skip import Skip
   rest/initializer.py:12:             from abjad.skip.skip import Skip
   skip/initializer.py:12:             from abjad.skip.skip import Skip
   tools/measuretools/make.py:2:       from abjad.skip.skip import Skip
   tools/measuretools/populate.py:4:   from abjad.skip.skip import Skip

These 'null' initializers can be made more useful.

We developed the tools/import package to automate the import of functions from the different tools packages. And we do that with 'active' initializers in the different tools packages. Here, for example, is the contents of tools/construct/__init__.py:

   from abjad.tools.imports.package_import import _package_import

   _package_import(__path__[0], globals( ))

It turns out that these same lines can be included in, for example, skip/__init__.py to remove one level of dot-reference to import typesetting classes:

   chord/initializer.py:24:            from abjad.skip import Skip
   metricgrid/format.py:4:             from abjad.skip import Skip
   note/initializer.py:13:             from abjad.skip import Skip
   rest/initializer.py:13:             from abjad.skip import Skip
   skip/initializer.py:13:             from abjad.skip import Skip
   tools/measuretools/make.py:3:       from abjad.skip import Skip
   tools/measuretools/populate.py:5:   from abjad.skip import Skip

Note that these imports of the form 'from abjad.skip import Skip' appear to opearate in exactly the same way as the longer imports of the form 'from abjad.skip.skip import Skip' that we have been using up to this point.

The commit here 'activates' only skip/__init__.py with the inclusion of 

   from abjad.tools.imports.package_import import _package_import

   _package_import(__path__[0], globals( ))

All tests pass.

Future commits will reproduce this pattern in as many of the other typesetting packages as possible.