Fixed 'format contention' bug in _ClefInterface.
Summary: Imagine a measure filled with notes. And then force a clef on the measure. Contention can then potentially arise between the measure and the first note in the measure. That is, both the measure and the first note in the measure might both realize that they should contribute a \clef statement at format time. We might then call this situation format-time contention between the measure and the first note in the measure, with both compoennts vying to be the component which contributes the \clef directive.

This same situation arose some months ago with _MeterInterface (which behaves very much the same as _ClefInterface in many ways). The solution there was to implement three private functions on _MeterInterface that find out whether the client of _MeterInterface *can* but then also *should* contribute a \time directive. Specifically, if the client of _MeterInterface *can* contribute a \time directive then the client must also check to make sure that no component in the parentage of the client can also contribute. When that's the case, the client of _MeterInterface *refrains from* contributing the \time directive, in deference to the component in the parentage that will contribute the \time directive.

This fix implements that exact same contention bypass logic in _ClefInterface.

New tests added.

Also, this means that the contention bypass logic needs to externalize from both _MeterInterface and _ClefInterface and live in some shared spot to avoid code duplication.