New regime to achieve absolutely fixed staff positioning through an entire score.
Three new layout objects::

   SystemYOffsets
   StaffAlignmentOffsets
   FixedStaffPositioning

The last of these three layout objects -- FixedStaffPositioning -- replaces the old FixedSystemsIndicator.

In addition, this commit renames layout.apply_fixed_systems_indicator( ) to layout.apply_fixed_staff_positioning( ). The new name is more appropriate because it is not just systems but staves (and systems) that this function sets.

Use is like this:

   t = Score(...)

   systems = SystemYOffsets(35, 7)
   staves = StaffAlignmentOffsets(0, -15)
   positioning = FixedStaffPositioning(systems, staves)

   layout.apply_fixed_staff_positioning(t, positioning)

That is:

   1. build some music
   2. instantiate both a SystemYOffsets and a StaffAlignmentOffsets object
   3. pass both of those into a new FixedStaffPositioning object
   4. pass the FixedStaffPositioning object into the layout function

The result is that LilyPond Score.NonMusicalPaperColumn Y-offset and alignment-offsets will apply correct throughout the entire score. Output is something like this:
  
            \overrideProperty #"Score.NonMusicalPaperColumn"
            #'line-break-system-details
            #'((Y-offset . 70) (alignment-offsets . (0 -15)))

Note that 70 is a multiple of 35. This means that the excerpt of .ly input given above represents the second system on the page. The layout given here is likely for, for example, piano music with two staves per system with systems spaced 35 page units apart running down the page.