New: measure-level barline overrides now populate the correct slots at format time to accommodate LilyPond barline creation procedures.
This is easiest to see from the new test file added with this commit.
This is easiest to see with an example.

   t = Staff(RigidMeasure((2, 8), construct.run(2)) * 3)
   pitchtools.diatonicize(t)
   t.formatter.number.measures = 'comment'

   t[0].barline.kind = '||'
   t[0].barline.color = 'red'
   t[0].barline.promote('color', 'Staff')

   t[0].meter.color = 'blue'
   t[0].meter.promote('color', 'Staff')

   \new Staff {
           % start measure 1
                   \override Staff.TimeSignature #'color = #blue
                   \time 2/8
                   c'8
                   d'8
                   \override Staff.BarLine #'color = #red
                   \bar "||"
                   \revert Staff.TimeSignature #'color
           % stop measure 1
           % start measure 2
                   \time 2/8
                   e'8
                   \revert Staff.BarLine #'color
                   f'8
           % stop measure 2
           % start measure 3
                   \time 2/8
                   g'8
                   a'8
           % stop measure 3
   }


Notice that the paired BarLine \override and \revert commands inhabit very 
different format slots than the TimeSignature \override and \revert commands.

The TimeSignature \override and \revert commands live where we would expect
them: at the beginning and end of the first measure, bracketing the supposed
start and stop of the measure.

But the BarLine \override and \revert commands live in weird places.
The BarLine \override comes *AT THE END* of the first measure.
The BarLine \revert comes *AFTER THE FIRST NOTE* of *THE FOLLOWING MEASURE*.

Despite the strange appearance of the LilyPond code, the printed PDF
produces exactly what you would expect from the Abjad code.