Changed: t.leaves now returns an immutable tuple instead of a mutable list.
FIXED: the seemingly minor change listed above resulted in the chance discovery of a longstanding bug container initialization. Details follow.

The previous implementation of Container._initializeMusic( ) was this::

   def _initializeMusic(self, music):
      '''Insert components in 'music' in container.
         Set parent of components in 'music' to container.'''
      from abjad.tools import check
      from abjad.tools import parenttools
      from abjad.tools.parenttools.switch import _switch
      music = music or [ ]
      check.assert_components(music, contiguity = 'strict', share = 'thread')
      parent, index, stop_index = parenttools.get_with_indices(music)
      self._music = music
      _switch(self._music, self)
      if parent is not None:
         parent._music.insert(index, self)
         self.parentage._switch(parent)

There was, however, a substantial problem in the following line:

      self._music = music

Why?

Because this line binds self._muisc to THE EXACT MUSIC LIST PASSED IN. That is, if the `music` list passed in for initialization had Python id 17621680 then, after initialization, id(container._music) would also equal exactly 17621680.

The solution is straightforward:

      self._music = music[:]

or

      self._music = list(music)

I chose the second form with the explicit list constructor because the input `music` list may now, as of this commit, also be a tuple of components rather than only a list of components. The music[:] form would therefore return only a newly constructed tuple (rather than a newly constructed) list. The list(music) form guarantees that self._music is a list after assignment.

The potential impact and weirdness of the old behavior can be seen easily in the following example::

   abjad> music = construct.scale(4)
   abjad> staff = Staff(music)
   abjad> id(music)
   17864504
   abjad> id(staff._music)
   17864504

   abjad> music.append(Rest((1, 4)))

   abjad> music
   [Note(c', 8), Note(d', 8), Note(e', 8), Note(f', 8), Rest(4)]

   abjad> f(staff)
   \new Staff {
           c'8
           d'8
           e'8
           f'8
           r4
   }

This is clearly a disaster.

With the fix committed here, this problem goes away::

   abjad> music = construct.scale(4)
   abjad> staff = Staff(music)
   abjad> id(music)
   17864504
   abjad> id(staff._music)
   18753656

   abjad> music.append(Rest((1, 4)))

   abjad> music
   [Note(c', 8), Note(d', 8), Note(e', 8), Note(f', 8), Rest(4)]

   abjad> f(staff)
   \new Staff {
           c'8
           d'8
           e'8
           f'8
   }

The rest of the files changed in this commit merely accommodate the fact that t.leaves now returns a tuple instead of a list. This means that expressions like

   foo = [Note(0, (1, 4))] + t.leaves

will no longer work. Instead you need something like

   foo = [Note(0, (1, 4))] + list(t.leaves)

or

   foo = [ ]
   foo.append(Note(0, (1, 4)))
   foo.extend(t.leaves)

which is probably best.