_StaffInterface is now a _GrobHandler too. Grob overrides are applied to the LilyPond *StaffSymbol*. Thus it is now possible to say::
   t = Staff(construct.scale(4))
   t.staff.color = 'red'
   f(t)
   \new Staff \with {
           \override StaffSymbol #'color = #red
   } {
           c'8
           d'8
           e'8
           f'8
   }

.. todo:: note that, while StaffSymbol attributes can be set from any Abjad component, setting from a component other than Staff will not necessarily render the changes as expected. Example::

   t = Staff(construct.scale(4))
   t[2].staff.color = 'red'
   t[2].staff.promote('color', 'Staff')
   f(t)

   \new Staff {
           c'8
           d'8
           \once \override Staff.StaffSymbol #'color = #red
           e'8
           f'8
   } 

This is because the LilyPond Staff must be stoped and restarted (via the \stopStaff\startStaff directives) before the new overrides are sent.
Should the GrobHandler be smart enough to see who is contributing the formating and add a \stopStaff\startStaff immediately before the staff contribution IFF the contributor is not a Staff? Or should the user be responsible for handling this?