New: tools/markup package.
Initial package contents:

   markup.big_centered_page_number( )


Description: the new tools/markup package establishes a space to implement public functions, like markup.big_centered_page_number( ), that are designed to return an Abjad Markup object filled out in some useful way.

For example:

   abjad> t = markup.big_centered_page_number( )
   abjad> print t.format
   \markup { 
      \fill-line {
      \bold \fontsize #3 \concat {
      \on-the-fly #print-page-number-check-first
      \fromproperty #'page:page-number-string } } }

The new tools/markup package is a package like any other, complete with a tools/markup/test directory and with API entries for the functions contained within.

Note that this does mean that 'markup' is now an Abjad-global term and should not be used as a variable name in code used to build individual scores.