Sphinx now builds the HTML API entries with zero warnings.
It was necessary to teach make_sphinx_module_listing.py about items like Crescendo and Decrescendo, which look like classes (with an initial capital letter) but are, in fact, functions that wrap some other class (Hairpin, in the case of Crescendo and Decrescendo).

The important point here is that the codebase *does* implement public functions *outside of* the tools/ packages. There are not many of these non-tool public functions. So few, in fact, that it might be worth making Crescendo and Decrescendo into full-fledged classes and then migrating navigator/dfs.py to tools/iterate. But for now the docs build correctly again.