Ran the following in abjad/documentation:
   svn propset svn:ignore _build .

This means that Subversion now ignores any files or directories in 
abjad/documentation beginning in the string '_build', including,
of course, the abjad/documentation/_build directory.

svn:ignore was set on 20 May to apply to abjad/documentation/doc_sphinx,
which has since been removed and replaced with abjad/documentation instead.