Changed Spanner.copy(start = None, stop = None) to Spanner.copy(components).
Previously, Spanner.copy(start = None, stop = None) worked by accepting
integer start and stop "indices". Note that "indices" is in scare quotes
because the interpretation of these indices did NOT match the interpretation
of index arguments to the built-in Python list, tuple and other iterable
types. What this meant was that private code internal to the implementation
of Spanner was calling Spanner.copy( ) with weird values passed in.

As of this commit, Spanner.copy(components) no longer accepts integer
values of any sort. Instead, Spanner.copy(components) now accepts any
iterable of components that are already contained in spanner.

Since only private code internal to the implementation of Spanner was
previously calling Spanner.copy( ), only private code internal to Spanner
changes in this commit.

This commit preserves Spanner.copy( ) in the public interface.

The next commit will remove Spanner.copy( ) from the public interface
because it is unneeded and unused in composer activities.