New: three new show( ) optional keywords:
   show(t, formattime = 10)
   show(t, return_timing = False)
   show(t, suppress_pdf = False)


You can set show(t, formattime = whatever) in similarly to
show(t, lilytime = whatever) to print the integer number of seconds
that it takes Abjad to format t.

   abjad> t = Staff(construct.run(500))
   abjad> show(t, formattime = 1)
   Abjad format time equal to 3 sec.

This satisfies Google enhancement #19.

You can set show(t, return_timing = True) to return a pair
of the integer number of seconds that it takes

   abjad> t = Staff(construct.run(500))
   abjad> show(t, return_timing = True)
   (2, 3)

And you can set show(t, suppress_pdf = True) to cause show to go
through the usual format-and-render process, but to not open
the PDF as the last step in the process. This is useful when
writing regression tests.

   t = Staff(construct.run(4))
   result = show(t, return_timing = True, suppress_pdf = True)
   assert isinstance(result, tuple)
   assert len(result) == 2
   assert isinstance(result[0], int)
   assert isinstance(result[1], int)

Because suppress_pdf is useful only when writing regression tests,
the keyword is not documented.

New test file added: tools/io/show/test_io_show_return_timing.py.
This is our first first io test file. It contains only a single
test case. But that test case takes almost 4 seconds to run
because the test case calls LilyPond behind the scenes.