New: scr/devel/fix-test-case-names developer script.
Run the new script in the root of your Abjad install directory.
Right now the script checks only for test case names that do not
match the name of the test file in which they are implemented.

   abjad$ scr/devel/fix-test-case-names 
   < ... many lines omitted ... >
   ONMATCHING: test_divide_pair.py test_divide_no_prolation_01
   NONMATCHING: test_divide_pair.py test_divide_no_prolation_02
   NONMATCHING: test_divide_pair.py test_divide_no_prolation_03
   NONMATCHING: test_divide_pair.py test_divide_no_prolation_04
   NONMATCHING: test_divide_pair.py test_lone_divide_01
   NONMATCHING: test_divide_pair.py test_lone_divide_02
   NONMATCHING: test_divide_pair.py test_lone_divide_03
   NONMATCHING: test_divide_pair.py test_lone_divide_04
   NONMATCHING: test_divide_pair.py test_no_empty_list_to_lone_divide
   NONMATCHING: test_divide_pair.py test_no_zero_value_to_lone_divide
   NONMATCHING: test_divide_pair.py test_no_zero_values_to_divide

   614 total test modules.
   2270 total test cases.
   498 total nonmatching test case name.

More documentation and functionality coming soon.

The other changes in this commit are changes to make test case names
match test file names in accordance with the output of this new script.