New: lilytools module now available.
This module implements eight new public classes:

   BookBlock
   BookpartBlock
   HeaderBlock
   LayoutBlock
   LilyFile
   MidiBlock
   PaperBlock
   ScoreBlock

The LilyFile class models the LilyPond .ly input file structure.

The remaining seven classes divide into blocks that handle user
attribute settings (HeaderBlock, LayoutBlock, MidiBlock, PaperBlock)
and blocks that don't (BookBlock, BookpartBlock, ScoreBlock).

EXAMPLE:

   score = Score([Staff(construct.scale(4))])

   header_block = lilytools.HeaderBlock( )
   header_block.composer = Markup('Josquin')
   header_block.title = Markup('Missa sexti toni')
   layout_block = lilytools.LayoutBlock( )
   layout_block.indent = 0
   layout_block.ragged_right = True
   paper_block = lilytools.PaperBlock( )
   paper_block.top_margin = 15
   paper_block.left_margin = 15
   midi_block = lilytools.MidiBlock( )

   score_block = lilytools.ScoreBlock( )
   score_block.append(score)
   score_block.append(layout_block)
   score_block.append(midi_block)

   lily_file = lilytools.LilyFile( )
   lily_file.append(paper_block)
   lily_file.append(score_block)

THEN:

   abjad> f(lily_file)

   \paper {
           left-margin = #15
           top-margin = #15
   }

   \score {
           \new Score <<
                   \new Staff {
                           c'8
                           d'8
                           e'8
                           f'8
                   }
           >>
           \layout {
                   indent = #0
                   ragged-right = ##t
           }
           \midi { }
   }

The lilytools components are, at the moment, all classes instead of functions.

This marks the first time we've put class constructors inside a tools module.
This is also currently breaking the docuentation/scr/make-abjad-api because
the logic inside that script assumes that tools modules contain only
functions. This isn't hurting anything and will probably stay broken for
a while. (Note that Sphinx will issue 8 build warnings for the time being.)

Note too that the lilytools components are intentionally relatively 'dumb'
as of this implementation. The goal is to introduce the components to the
system and use them for a while. Then later we can migrate functionality
to LilyFile, including the insertion of include files and the setting
of Scheme functions like set-global-staff-size and friends.